mods/QRTools/requirements.txt
scipy
qrcode
gensim
Pillow
pyzbar
opencv-python
